chore(package.json): remove predeploy script
The predeploy script was removed as it is no longer needed. It was previously used to rename the main.js and main.css files in the build/static/js and build/static/css directories respectively. This step is no longer necessary and can be handled by the build process itself.